requirements.txt
google-cloud-aiplatform>=1.71.0
google-cloud-discoveryengine>=0.13.0
google-cloud-storage>=2.0.0
vertexai>=1.71.0
beir>=2.0.0
pytrec-eval-terrier>=0.5.6
pandas>=2.0.0
tabulate>=0.9.0
python-dotenv>=1.0.0
tqdm>=4.66.0
